reorganise files and components
