+++ frontend-two/pages/on-time/[nocCode].tsx
@@ -23,7 +23,7 @@ import {
   DISPLAY_MODE_OPTIONS,
   normaliseDirection,
   aggregatePerformanceTotals,
-} from "@/utils/on-time-table-format";
+} from "@/utils/on-time/on-time-table-format";
 import {
   refineResultsToPerformanceFilters,
   performanceFiltersToRefineResults,
